# Unroll the ChaCha20 inner loop for performance