@kristapsk 2022-04-22
Concept ACK, I see `./src/bench/bench_bitcoin` improvements with this change.