@sipa 2022-04-22
@instagibbs 

`./src/bench/bench_bitcoin`