@laanwj 2022-04-22
I'm somewhat surprised unrolling a loop that is 20 times the same thing gives that much performance win on modern CPUs. But it's just a few ROR instrutions I guess so the loop overhead easily dominates? Anyhow, concept ACK.