@instagibbs 2022-04-22
Getting a rough average of 15% speedup as well