@laanwj 2022-04-22
Gcc 11.2.0, x86_64:
- The function `ChaCha20::Keystream` grows in size from 992 bytes to 3840 (doesn't seem too bad, still fits in a page).
- One iteration of the loop looks like:
```
 370:	41 01 ed             	add    %ebp,%r13d
 373:	41 01 db             	add    %ebx,%r11d
 376:	41 01 f2             	add    %esi,%r10d
 379:	44 31 e9             	xor    %r13d,%ecx
 37c:	44 31 da             	xor    %r11d,%edx
 37f:	44 31 d0             	xor    %r10d,%eax
 382:	c1 c1 10             	rol    $0x10,%ecx
 385:	c1 c2 10             	rol    $0x10,%edx
 388:	41 01 c9             	add    %ecx,%r9d
 38b:	01 d7                	add    %edx,%edi
 38d:	c1 c0 10             	rol    $0x10,%eax
 390:	44 31 cd             	xor    %r9d,%ebp
 393:	31 fb                	xor    %edi,%ebx
 395:	41 01 c4             	add    %eax,%r12d
 398:	c1 c5 0c             	rol    $0xc,%ebp
 39b:	c1 c3 0c             	rol    $0xc,%ebx
 39e:	44 31 e6             	xor    %r12d,%esi
 3a1:	41 01 ed             	add    %ebp,%r13d
 3a4:	41 01 db             	add    %ebx,%r11d
 3a7:	c1 c6 0c             	rol    $0xc,%esi
 3aa:	44 31 e9             	xor    %r13d,%ecx
 3ad:	44 31 da             	xor    %r11d,%edx
 3b0:	41 01 f2             	add    %esi,%r10d
 3b3:	c1 c1 08             	rol    $0x8,%ecx
 3b6:	c1 c2 08             	rol    $0x8,%edx
 3b9:	44 31 d0             	xor    %r10d,%eax
 3bc:	41 01 c9             	add    %ecx,%r9d
 3bf:	01 d7                	add    %edx,%edi
 3c1:	44 31 cd             	xor    %r9d,%ebp
 3c4:	31 fb                	xor    %edi,%ebx
 3c6:	89 7c 24 08          	mov    %edi,0x8(%rsp)
 3ca:	c1 c5 07             	rol    $0x7,%ebp
 3cd:	c1 c3 07             	rol    $0x7,%ebx
 3d0:	44 89 4c 24 04       	mov    %r9d,0x4(%rsp)
 3d5:	c1 c0 08             	rol    $0x8,%eax
 3d8:	45 01 f8             	add    %r15d,%r8d
 3db:	41 01 dd             	add    %ebx,%r13d
 3de:	45 31 c6             	xor    %r8d,%r14d
 3e1:	41 01 c4             	add    %eax,%r12d
 3e4:	44 89 f7             	mov    %r14d,%edi
 3e7:	44 8b 74 24 0c       	mov    0xc(%rsp),%r14d
 3ec:	44 31 e6             	xor    %r12d,%esi
 3ef:	c1 c7 10             	rol    $0x10,%edi
 3f2:	c1 c6 07             	rol    $0x7,%esi
 3f5:	41 01 fe             	add    %edi,%r14d
 3f8:	41 01 f3             	add    %esi,%r11d
 3fb:	45 31 f7             	xor    %r14d,%r15d
 3fe:	45 89 f1             	mov    %r14d,%r9d
 401:	44 31 d9             	xor    %r11d,%ecx
 404:	41 c1 c7 0c          	rol    $0xc,%r15d
 408:	c1 c1 10             	rol    $0x10,%ecx
 40b:	45 01 f8             	add    %r15d,%r8d
 40e:	44 31 c7             	xor    %r8d,%edi
 411:	c1 c7 08             	rol    $0x8,%edi
 414:	41 01 f9             	add    %edi,%r9d
 417:	44 31 ef             	xor    %r13d,%edi
 41a:	c1 c7 10             	rol    $0x10,%edi
 41d:	45 31 cf             	xor    %r9d,%r15d
 420:	41 01 c9             	add    %ecx,%r9d
 423:	41 01 fc             	add    %edi,%r12d
 426:	41 c1 c7 07          	rol    $0x7,%r15d
 42a:	44 31 e3             	xor    %r12d,%ebx
 42d:	c1 c3 0c             	rol    $0xc,%ebx
 430:	41 01 dd             	add    %ebx,%r13d
 433:	44 31 ef             	xor    %r13d,%edi
 436:	41 89 fe             	mov    %edi,%r14d
 439:	41 c1 c6 08          	rol    $0x8,%r14d
 43d:	45 01 f4             	add    %r14d,%r12d
 440:	44 31 e3             	xor    %r12d,%ebx
 443:	c1 c3 07             	rol    $0x7,%ebx
 446:	44 31 ce             	xor    %r9d,%esi
 449:	45 01 fa             	add    %r15d,%r10d
 44c:	41 01 e8             	add    %ebp,%r8d
 44f:	c1 c6 0c             	rol    $0xc,%esi
 452:	44 31 d2             	xor    %r10d,%edx
 455:	44 31 c0             	xor    %r8d,%eax
 458:	c1 c2 10             	rol    $0x10,%edx
 45b:	41 01 f3             	add    %esi,%r11d
 45e:	c1 c0 10             	rol    $0x10,%eax
 461:	44 31 d9             	xor    %r11d,%ecx
 464:	c1 c1 08             	rol    $0x8,%ecx
 467:	41 8d 3c 09          	lea    (%r9,%rcx,1),%edi
 46b:	44 8b 4c 24 04       	mov    0x4(%rsp),%r9d
 470:	31 fe                	xor    %edi,%esi
 472:	89 7c 24 0c          	mov    %edi,0xc(%rsp)
 476:	8b 7c 24 08          	mov    0x8(%rsp),%edi
 47a:	41 01 d1             	add    %edx,%r9d
 47d:	c1 c6 07             	rol    $0x7,%esi
 480:	01 c7                	add    %eax,%edi
 482:	45 31 cf             	xor    %r9d,%r15d
 485:	31 fd                	xor    %edi,%ebp
 487:	41 c1 c7 0c          	rol    $0xc,%r15d
 48b:	c1 c5 0c             	rol    $0xc,%ebp
 48e:	45 01 fa             	add    %r15d,%r10d
 491:	41 01 e8             	add    %ebp,%r8d
 494:	44 31 d2             	xor    %r10d,%edx
 497:	44 31 c0             	xor    %r8d,%eax
 49a:	c1 c2 08             	rol    $0x8,%edx
 49d:	c1 c0 08             	rol    $0x8,%eax
 4a0:	41 01 d1             	add    %edx,%r9d
 4a3:	01 c7                	add    %eax,%edi
 4a5:	45 31 cf             	xor    %r9d,%r15d
 4a8:	31 fd                	xor    %edi,%ebp
 4aa:	41 c1 c7 07          	rol    $0x7,%r15d
 4ae:	c1 c5 07             	rol    $0x7,%ebp
 4b1:	83 6c 24 10 01       	subl   $0x1,0x10(%rsp)
 4b6:	0f 85 b4 fe ff ff    	jne    370 <ChaCha20::Keystream(unsigned char*, unsigned long)+0x140>
```
- The unrolling indeed causes different register allocation, as well as instructions from multiple iterations to be interspersed (maybe better for scheduling, maybe it's possible to combine?).
- Benchmarks before on old AMD Phenom(tm) II X6 1075T:
```
|             ns/byte |              byte/s |    err% |     total | benchmark
|--------------------:|--------------------:|--------:|----------:|:----------
|                2.18 |      459,187,395.75 |    0.3% |      0.03 | `CHACHA20_1MB`
|                2.21 |      452,155,530.63 |    0.2% |      0.01 | `CHACHA20_256BYTES`
|                2.34 |      427,257,435.31 |    0.0% |      0.01 | `CHACHA20_64BYTES`
```
- Benchmarks after on same (~12% speedup):
```
|             ns/byte |              byte/s |    err% |     total | benchmark
|--------------------:|--------------------:|--------:|----------:|:----------
|                1.91 |      523,324,820.67 |    0.4% |      0.02 | `CHACHA20_1MB`
|                1.94 |      516,638,576.63 |    0.0% |      0.01 | `CHACHA20_256BYTES`
|                2.22 |      451,258,216.13 |    4.6% |      0.01 | `CHACHA20_64BYTES`
```